DDF-04481 Combined the BasicAuthenticationHandler and the WssBasicAuthenticationHandler.
