+++ platform/security/handler/security-handler-basic/src/test/java/org/codice/ddf/security/handler/basic/BasicAuthenticationHandlerTest.java
@@ -22,25 +22,34 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.when;
 
+import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import java.util.Base64;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.ws.rs.core.HttpHeaders;
-import org.codice.ddf.parser.Parser;
-import org.codice.ddf.parser.xml.XmlParser;
+import javax.xml.parsers.DocumentBuilder;
+import javax.xml.parsers.DocumentBuilderFactory;
+import javax.xml.parsers.ParserConfigurationException;
 import org.codice.ddf.platform.filter.FilterChain;
 import org.codice.ddf.security.handler.api.BaseAuthenticationToken;
 import org.codice.ddf.security.handler.api.HandlerResult;
-import org.codice.ddf.security.handler.api.UPAuthenticationToken;
 import org.junit.Test;
 import org.mockito.Mockito;
+import org.w3c.dom.Document;
+import org.w3c.dom.Node;
+import org.w3c.dom.NodeList;
+import org.xml.sax.SAXException;
 
 public class BasicAuthenticationHandlerTest {
 
   private static final String USERNAME = "admin";
 
   private static final String PASSWORD = "password";
+
+  private static final String USERNAME_ATTR = "Username";
+
+  private static final String PASSWORD_ATTR = "Password";
 
   private static final String CREDENTIALS = USERNAME + ":" + PASSWORD;
 
@@ -71,7 +80,7 @@ public class BasicAuthenticationHandlerTest {
    * flag is set) - both requests without and with the credentials are tested.
    */
   @Test
-  public void testGetNormalizedTokenResolveWithCredentials() throws IOException {
+  public void testGetNormalizedTokenResolveWithCredentials() throws Exception {
     BasicAuthenticationHandler handler = new BasicAuthenticationHandler();
 
     HttpServletRequest request = mock(HttpServletRequest.class);
@@ -85,8 +94,8 @@ public class BasicAuthenticationHandlerTest {
 
     assertNotNull(result);
     assertEquals(HandlerResult.Status.COMPLETED, result.getStatus());
-    assertEquals("admin", result.getToken().getPrincipal());
-    assertEquals("password", result.getToken().getCredentials());
+    assertEquals("admin", getXmlAttributeValue(result.getToken(), USERNAME_ATTR));
+    assertEquals("password", getXmlAttributeValue(result.getToken(), PASSWORD_ATTR));
 
     // confirm that no responses were sent through the HttpResponse
     Mockito.verify(response, never()).setHeader(anyString(), anyString());
@@ -100,7 +109,7 @@ public class BasicAuthenticationHandlerTest {
    * flag is set) and UsernameTokenType was created from the HTTP request.
    */
   @Test
-  public void testGetNormalizedTokenResolveCompleted() {
+  public void testGetNormalizedTokenResolveCompleted() throws Exception {
     BasicAuthenticationHandler handler = new BasicAuthenticationHandler();
 
     HttpServletRequest request = mock(HttpServletRequest.class);
@@ -114,25 +123,7 @@ public class BasicAuthenticationHandlerTest {
 
     assertNotNull(result);
     assertEquals(HandlerResult.Status.COMPLETED, result.getStatus());
-    assertEquals("admin", result.getToken().getPrincipal());
-  }
-
-  /**
-   * This test case handles the scenario in which the credentials are not to be obtained (i.e.
-   * resolve flag is not set) and the UsernameTokenType could not be created with the HTTP request.
-   */
-  @Test
-  public void testGetNormalizedTokenNoResolveNoAction() {
-    BasicAuthenticationHandler handler = new BasicAuthenticationHandler();
-
-    HttpServletRequest request = mock(HttpServletRequest.class);
-    HttpServletResponse response = mock(HttpServletResponse.class);
-    FilterChain chain = mock(FilterChain.class);
-
-    HandlerResult result = handler.getNormalizedToken(request, response, chain, false);
-
-    assertNotNull(result);
-    assertEquals(HandlerResult.Status.NO_ACTION, result.getStatus());
+    assertEquals("admin", getXmlAttributeValue(result.getToken(), USERNAME_ATTR));
   }
 
   /**
@@ -141,7 +132,7 @@ public class BasicAuthenticationHandlerTest {
    * request.
    */
   @Test
-  public void testGetNormalizedTokenNoResolveCompleted() {
+  public void testGetNormalizedTokenNoResolveCompleted() throws Exception {
     BasicAuthenticationHandler handler = new BasicAuthenticationHandler();
 
     HttpServletRequest request = mock(HttpServletRequest.class);
@@ -155,57 +146,49 @@ public class BasicAuthenticationHandlerTest {
 
     assertNotNull(result);
     assertEquals(HandlerResult.Status.COMPLETED, result.getStatus());
-    assertEquals("admin", result.getToken().getPrincipal());
+    assertEquals("admin", getXmlAttributeValue(result.getToken(), USERNAME_ATTR));
   }
 
   @Test
-  public void testExtractAuthInfo() {
-    Parser parser = new XmlParser();
+  public void testExtractAuthInfo() throws Exception {
     BasicAuthenticationHandler handler = new BasicAuthenticationHandler();
-    UPAuthenticationToken result =
-        (UPAuthenticationToken)
-            handler.extractAuthInfo(
-                "Basic " + Base64.getEncoder().encodeToString(CREDENTIALS.getBytes()));
-    assertNotNull(result);
-    assertEquals("admin", result.getUsername());
-    assertEquals("password", result.getPassword());
-
-    WssBasicAuthenticationHandler wssHandler = new WssBasicAuthenticationHandler(parser);
-    BaseAuthenticationToken wssResult =
-        wssHandler.extractAuthInfo(
+    BaseAuthenticationToken result =
+        handler.extractAuthInfo(
             "Basic " + Base64.getEncoder().encodeToString(CREDENTIALS.getBytes()));
-    assertNotNull(wssResult);
-
-    result =
-        (UPAuthenticationToken)
-            handler.extractAuthInfo(
-                "Basic " + Base64.getEncoder().encodeToString(":password".getBytes()));
     assertNotNull(result);
-    assertEquals("", result.getUsername());
-    assertEquals("password", result.getPassword());
+    assertEquals("admin", getXmlAttributeValue(result, USERNAME_ATTR));
+    assertEquals("password", getXmlAttributeValue(result, PASSWORD_ATTR));
 
+    handler = new BasicAuthenticationHandler();
     result =
-        (UPAuthenticationToken)
-            handler.extractAuthInfo(
-                "Basic " + Base64.getEncoder().encodeToString("user:".getBytes()));
+        handler.extractAuthInfo(
+            "Basic " + Base64.getEncoder().encodeToString(CREDENTIALS.getBytes()));
     assertNotNull(result);
-    assertEquals("user", result.getUsername());
-    assertEquals("", result.getPassword());
 
     result =
-        (UPAuthenticationToken)
-            handler.extractAuthInfo(
-                "Basic " + Base64.getEncoder().encodeToString("user/password".getBytes()));
+        handler.extractAuthInfo(
+            "Basic " + Base64.getEncoder().encodeToString(":password".getBytes()));
+    assertNotNull(result);
+    assertEquals("", getXmlAttributeValue(result, USERNAME_ATTR));
+    assertEquals("password", getXmlAttributeValue(result, PASSWORD_ATTR));
+
+    result =
+        handler.extractAuthInfo("Basic " + Base64.getEncoder().encodeToString("user:".getBytes()));
+    assertNotNull(result);
+    assertEquals("user", getXmlAttributeValue(result, USERNAME_ATTR));
+    assertEquals("", getXmlAttributeValue(result, PASSWORD_ATTR));
+
+    result =
+        handler.extractAuthInfo(
+            "Basic " + Base64.getEncoder().encodeToString("user/password".getBytes()));
     assertNull(result);
 
-    result =
-        (UPAuthenticationToken)
-            handler.extractAuthInfo("Basic " + Base64.getEncoder().encodeToString("".getBytes()));
+    result = handler.extractAuthInfo("Basic " + Base64.getEncoder().encodeToString("".getBytes()));
     assertNull(result);
   }
 
   @Test
-  public void testExtractAuthenticationInfo() {
+  public void testExtractAuthenticationInfo() throws Exception {
     // only test valid authorization header and missing header - invalid values are tested in
     // textExtractAuthInfo
     BasicAuthenticationHandler handler = new BasicAuthenticationHandler();
@@ -214,35 +197,58 @@ public class BasicAuthenticationHandlerTest {
     when(request.getHeader(HttpHeaders.AUTHORIZATION))
         .thenReturn("Basic " + Base64.getEncoder().encodeToString(CREDENTIALS.getBytes()));
 
-    UPAuthenticationToken result =
-        (UPAuthenticationToken) handler.extractAuthenticationInfo(request);
+    BaseAuthenticationToken result = handler.extractAuthenticationInfo(request);
     assertNotNull(result);
-    assertEquals("admin", result.getUsername());
-    assertEquals("password", result.getPassword());
+    assertEquals("admin", getXmlAttributeValue(result, USERNAME_ATTR));
+    assertEquals("password", getXmlAttributeValue(result, PASSWORD_ATTR));
 
-    result = (UPAuthenticationToken) handler.extractAuthenticationInfo(request);
+    result = handler.extractAuthenticationInfo(request);
     assertNotNull(result);
-    assertEquals("admin", result.getUsername());
-    assertEquals("password", result.getPassword());
+    assertEquals("admin", getXmlAttributeValue(result, USERNAME_ATTR));
+    assertEquals("password", getXmlAttributeValue(result, PASSWORD_ATTR));
 
     when(request.getHeader(HttpHeaders.AUTHORIZATION)).thenReturn(null);
-    result = (UPAuthenticationToken) handler.extractAuthenticationInfo(request);
+    result = handler.extractAuthenticationInfo(request);
     assertNull(result);
   }
 
   @Test(expected = IllegalStateException.class)
-  public void testIllegalStateException() {
+  public void testIllegalStateException()
+      throws ParserConfigurationException, IOException, SAXException {
     BasicAuthenticationHandler handler = new BasicAuthenticationHandler();
-    UPAuthenticationToken result =
-        (UPAuthenticationToken)
-            handler.extractAuthInfo(
-                "Basic " + Base64.getEncoder().encodeToString(CREDENTIALS.getBytes()));
+    BaseAuthenticationToken result =
+        handler.extractAuthInfo(
+            "Basic " + Base64.getEncoder().encodeToString(CREDENTIALS.getBytes()));
     assertNotNull(result);
-    assertEquals("admin", result.getUsername());
-    assertEquals("password", result.getPassword());
+    assertEquals("admin", getXmlAttributeValue(result, USERNAME_ATTR));
+    assertEquals("password", getXmlAttributeValue(result, PASSWORD_ATTR));
 
-    WssBasicAuthenticationHandler wssHandler = new WssBasicAuthenticationHandler(null);
-    wssHandler.extractAuthInfo(
+    BasicAuthenticationHandler basicHandler = new BasicAuthenticationHandler();
+    basicHandler.extractAuthInfo(
         "Basic " + Base64.getEncoder().encodeToString(CREDENTIALS.getBytes()));
+  }
+
+  private String getXmlAttributeValue(BaseAuthenticationToken token, String attribute)
+      throws ParserConfigurationException, IOException, SAXException {
+    DocumentBuilderFactory factory = DocumentBuilderFactory.newInstance();
+    DocumentBuilder builder = factory.newDocumentBuilder();
+    ByteArrayInputStream input =
+        new ByteArrayInputStream(token.getCredentialsAsXMLString().getBytes("UTF-8"));
+    Document doc = builder.parse(input);
+    NodeList children = doc.getDocumentElement().getChildNodes();
+
+    int childIndex = 0;
+    while (childIndex < children.getLength()) {
+      Node child = children.item(childIndex);
+      if (child.getNodeName().contains(attribute)) {
+        if (child.getFirstChild() != null) {
+          return child.getFirstChild().getNodeValue();
+        } else {
+          return "";
+        }
+      }
+      childIndex++;
+    }
+    return "";
   }
 }
